Moved interface adapter elements for choose playlists into a separate package
