+++ src/interface_adapter/choose_playlist/ChoosePlaylistState.java
@@ -1,4 +1,4 @@
-package interface_adapter;
+package interface_adapter.choose_playlist;
 
 public class ChoosePlaylistState {
 }
