+++ src/interface_adapter/choose_playlist/ChoosePlaylistViewModel.java
@@ -1,4 +1,4 @@
-package interface_adapter;
+package interface_adapter.choose_playlist;
 
 public class ChoosePlaylistViewModel {
 }
